+++ src/utils/const/inputFields.js
@@ -1,0 +1,2 @@
+export const TASK_TEXTAREA_NUM_OF_ROW = 4;
+export const TEXTAREA_DEFAULT_ROW = 4;
